Code cleanup: separate TriggerArea from collider
