+++ src/Nodes/CMakeLists.txt
@@ -5,6 +5,7 @@ set(SOURCES
     FPSCounter.cpp
     Colliders.cpp
     CollisionManager.cpp
+    TriggerArea.cpp
 )
 
 add_library(Nodes ${SOURCES})
